:recycle: (Venue) Extract top container from VenueBody
